汐りんのセリフ生成をカスタムフック化
